Finders: Remove duplcate finder code
