+++ demosys/core/finders.py
@@ -1,11 +1,13 @@
 """
 Base finders
 """
+import functools
 from collections import namedtuple
 from pathlib import Path
 
 from demosys.conf import settings
 from demosys.core.exceptions import ImproperlyConfigured
+from demosys.utils.module_loading import import_string
 
 FinderEntry = namedtuple('FinderEntry', ['path', 'abspath', 'exists'])
 
@@ -57,3 +59,19 @@ class BaseEffectDirectoriesFinder(BaseFileSystemFinder):
     def find(self, path: Path):
         path = Path(self.directory) / Path(path)
         return super().find(path)
+
+
+@functools.lru_cache(maxsize=None)
+def get_finder(import_path):
+    """
+    Get a finder class from an import path.
+    Raises ``demosys.core.exceptions.ImproperlyConfigured`` if the finder is not found.
+    This function uses an lru cache.
+
+    :param import_path: string representing an import path
+    :return: An instance of the finder
+    """
+    Finder = import_string(import_path)
+    if not issubclass(Finder, BaseFileSystemFinder):
+        raise ImproperlyConfigured('Finder {} is not a subclass of core.finders.FileSystemFinder'.format(import_path))
+    return Finder()
