+++ demosys/core/texturefiles/finders.py
@@ -1,9 +1,5 @@
-import functools
-
 from demosys.core import finders
 from demosys.conf import settings
-from demosys.core.exceptions import ImproperlyConfigured
-from demosys.utils.module_loading import import_string
 
 
 class FileSystemFinder(finders.BaseFileSystemFinder):
@@ -18,20 +14,4 @@ class EffectDirectoriesFinder(finders.BaseEffectDirectoriesFinder):
 
 def get_finders():
     for finder in settings.TEXTURE_FINDERS:
-        yield get_finder(finder)
-
-
-@functools.lru_cache(maxsize=None)
-def get_finder(import_path):
-    """
-    Get a finder class from an import path.
-    Raises ``demosys.core.exceptions.ImproperlyConfigured`` if the finder is not found.
-    This function uses an lru cache.
-
-    :param import_path: string representing an import path
-    :return: An instance of the finder
-    """
-    Finder = import_string(import_path)
-    if not issubclass(Finder, finders.BaseFileSystemFinder):
-        raise ImproperlyConfigured('Finder {} is not a subclass of core.finders.FileSystemFinder'.format(import_path))
-    return Finder()
+        yield finders.get_finder(finder)
